New Update Add function get_list

+++ packarch.py
@@ -7,36 +7,44 @@ from time import sleep as sl
 
 current_directoy = getcwd()
 
-hide = [">/dev/null 2>&1", "|ls > .out && rm -rf .out"]
-sinc = " -S "
+hide = [">/dev/null 2>&1", "|ls > .out && rm -rf .out", "clear"]
 sp = " "
+All_text_cond = True
+words = [
+    "Installing ",  # 0
+    "Installed",  # 1
+    "Update Complete",  # 2
+    "Updating repos",  # 3
+    "...",  # 4
+    "Updating",  # 5
+    "Searching",  # 6
+]
 
 
 def clear():
-    sys("clear")
-
-
-def installed():
-    print("Installed")
-    sl(1)
+    sys(hide[2])
 
 
 class verbose:
     def all():
-        hide.pop(1)
-        hide.pop(0)
-        hide.append("")
-        hide.append("")
+        hide[0] = ""
+        hide[1] = ""
+        hide[2] = ""
 
     def aur():
-        hide.pop(1)
-        hide.pop(0)
-        hide.append("")
-        hide.append("|ls > .out && rm -rf .out")
+        hide[2] = ""
+        hide[0] = ""
 
     def pacman():
-        hide.pop(1)
-        hide.append("")
+        hide[1] = ""
+        hide[2] = ""
+
+    def quit_clear():
+        hide[2] = ""
+
+
+def installed():
+    print(words[1])
 
 
 def check(nombre_check):
@@ -57,12 +65,12 @@ def check(nombre_check):
 def install(nombre_pacman, cond_1="", cond_2=""):
     match cond_1:
         case "-v":
-            print("Installing " + nombre_pacman + "...")
+            print(words[0] + nombre_pacman + words[4])
             sys("sudo pacman -S " + nombre_pacman + sp + cond_2 + sp + " --noconfirm")
-            print("Installed")
+            print(words[1])
         case _:
             clear()
-            print("Installing " + nombre_pacman + "...")
+            print(words[0] + nombre_pacman + words[4])
             sys(
                 "sudo pacman -S "
                 + nombre_pacman
@@ -77,33 +85,39 @@ def install(nombre_pacman, cond_1="", cond_2=""):
 
 def refresh():
     clear()
-    print("Updating repos...")
+    print(words[3], words[4])
     sys("sudo pacman -Syy " + hide[0])
     clear()
-    print("Repos Updated")
+    print(words[2])
 
 
 def remove(removes, cond_1=""):
-    match cond_1:
-        case "-v":
-            sys("sudo pacman -R" + sp + removes + " --noconfirm")
-        case "-f":
+    if check(removes) == True:
+        if "v" in cond_1:
+            pass
+            hide[0] = ""
+        if "f" in cond_1:
             sys("sudo pacman -Rdd" + sp + removes + " --noconfirm" + hide[0])
-        case _:
+        else:
             sys("sudo pacman -R" + sp + removes + " --noconfirm" + hide[0])
+    else:
+        print("El o los paquetes no se encontraron en su totalidad")
+    hide[0] = ">/dev/null 2>&1"
 
 
 class aur:
-    def install(apps_aur):
+    def install(apps_aur, cond_1=""):
+        if "h" in cond_1:
+            words[0, 4] = ""
         for i in apps_aur.split():
             clear()
             url = "https://aur.archlinux.org/" + i + ".git"
             chdir("/tmp")
-            print("Cloning " + i + "...")
+            print("Cloning " + i + words[4])
             sys("git clone " + url + hide[0])
             clear()
             chdir(i)
-            print("Installing " + i + "...")
+            print(words[0] + i + words[4])
             sys("makepkg -si --noconfirm" + hide[0])
             clear()
             chdir(current_directoy)
@@ -112,25 +126,25 @@ class aur:
     def manager(nombre_gestor, app_gestor, cond_1="", cond_2=""):
         match cond_1:
             case "-v":
-                print("Installing " + app_gestor + "...")
+                print(words[0] + app_gestor + words[4])
                 sys(
                     nombre_gestor
-                    + sinc
+                    + " -S"
                     + app_gestor
                     + sp
                     + cond_2
                     + sp
                     + " --noconfirm"
                 )
-                print("Installed")
+                print(words[1])
             case "-s":
                 sys(nombre_gestor + sp + app_gestor + hide[0])
             case _:
                 clear()
-                print("Installing " + app_gestor + "...")
+                print(words[0] + app_gestor + words[4])
                 sys(
                     nombre_gestor
-                    + sinc
+                    + " -S"
                     + app_gestor
                     + " --noconfirm"
                     + sp
@@ -144,11 +158,29 @@ class aur:
 def upgrade(condu_1="", condu_2=""):
     match condu_1:
         case "-v":
-            print("Updating...")
+            print(words[5], words[4])
             sys("sudo pacman -Syu --noconfirm" + sp + condu_2)
         case _:
             clear()
-            print("Actualizando...")
+            print(words[5], words[4])
             sys("sudo pacman -Syu --noconfirm" + sp + condu_1 + sp + hide[0])
             clear()
-            print("Update Complete")
+            print(words[2])
+
+
+def get_list(list, cond1=""):
+    if cond1 != "":
+        options = [""]
+        print(words[6], words[4])
+        if "o" in cond1:
+            options[0] = "Ss"
+            if "e" in cond1:
+                options[0] = "Si"
+        else:
+            if "l" in cond1:
+                options[0] = "Qs"
+                if "e" in cond1:
+                    options[0] = "Q"
+        sys("pacman -" + options[0] + sp + list)
+    else:
+        sys("pacman -Ss " + list)
